pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "graph"
version = "0.1.0"
description = "GraphQL API Gateway"
authors = ["Kirill Sumorokov <sumorokov.k@gmail.com>"]
packages = [{ include = "graph", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
aio-pika = "6.8.0"
aiojobs = "1.0.0"
fastapi = "0.70.0"
starlette-prometheus = "0.8.0"
strawberry-graphql = {version = "0.93.3", extras = ["fastapi"]}
structlog = "21.5.0"
ujson = "4.3.0"
uvicorn = {extras = ["standard"], version = "0.16.0"}
aiohttp = {extras = ["speedups"], version = "^3.8.1"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
coverage = "6.2"
faker = "10.0.0"
flake8 = "4.0.1"
flake8-aaa = "0.12.0"
flake8-bugbear = "21.11.29"
flake8-builtins-unleashed = "1.3.1"
flake8-comprehensions = "3.7.0"
flake8-import-order = "0.18.1"
flake8-print = "4.0.0"
flake8-pytest = "1.3"
ipython = "7.30.1"
mypy = "0.910"
pytest = "6.2.5"
rope = "0.22.0"
types-setuptools = "57.4.4"
types-ujson = "4.2.0"

[tool.mypy]
python_version = "3.10"

plugins = ["strawberry.ext.mypy_plugin", "pydantic.mypy"]

warn_return_any = true
warn_unused_configs = true
